feat(front-kit): cleanup, updated uikit.md
