+++ packages/falcon-front-kit/src/Locale/index.ts
@@ -1,1 +1,3 @@
 export * from './LocaleContext';
+export * from './dateTimeFormat';
+export * from './priceFormat';
